Move knock code to new file
You're telling me that "just worked!?"
